Implement trace replay driver for Dimemas
In order to make our interface work with Dimemas, we need a single
socket connection to Dimemas.  The IB packet generator in the existing
IB model assumes that each app can enqueue a maximum of one message.
To model this properly, we treat each "app" as an instance of an MPI
rank which queues messages to the packet generator.

This commit adds a dispatcher to wait for Dimemas commands and inject
messages into the IB fabric.  When the message reaches its destination,
the Sink will notify the driver so that Dimemas can react accordingly.
The trace replay driver keeps all metadata for each message.  It is
important that we keep the metadata so that we can echo back the data
properly to Dimemas in the "SEND COMPLETED" message.  This is
accomplished by giving every message from the controller a unique ID
that replaces the application-specific ID.

Synchronization with Dimemas is done by running each simulation only
when the other is blocked on the socket wait.  Dimemas uses STOP
messages to accomplish this.  We schedule each internal STOP
self-message with the lowest possible priority, so that we do everything
we possibly can before returning control back to Dimemas.

The replay driver takes three parameters: (i) a parameter to scale
Dimemas timestamps according to the trace being replayed, (ii) the
IPv4 address to listen on, and (iii) the TCP port to listen on.
Dimemas does not use getaddrinfo() or an equivalent, so it can only
connect to dotted decimal IPv4 addresses.

Since Dimemas uses floating-point timestamps, it is possible for drift
due to rounding error to occur [1].  We need to fix this up if we notice
it to prevent bizarre behavior from OMNet++.

Dimemas inserts a tag into the SEND message, but does not expect to
receive it in the COMPLETED SEND message.  How's that for confusing?

By default, OMNet++ forces all gates to be connected.  However, we still
want to be able to run the model with synthetic probabilistic traffic
patterns.  Thus, we use the @loose annotation to tell OMNet++ that the
gates can be unconnected, and test for this before sending messages
through the "done" gate (we don't need to test for the input gate since
we simply won't get messages on it if it's unconnected).

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>