Add socket interface for Dimemas communication
We do socket I/O through C stdio, since this allows us to do I/O one
line at a time and take advantage of OS line buffering, rather than
implementing it ourselves.

We must check for exceptions in sendLine and we must only throw an
exception in recvLine if there was actually an error.  EOF is not
exceptional and should be handled by caller normally.

The standard library interfaces rely on seeking to synchronize their
buffer on read() and write() operations.  This doesn't work with
sockets, so the easiest way to fix this is to open one file stream for
reading and one for writing, avoiding this conflict entirely.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>