Change type of flitSize from double to int
The flitSize defined by the packet generator is int, but it was double
in the Sink.  This caused the OMNet++ IDE to report an error in the
modules.ini configuration file since flitSize was defined once but the
types were different.  This commit changes the type to be consistently
be int, since having a not-integral number of bytes does not make any
sense semantically.

This "bug" was present in the initial version of the simulation.

This patch does not change the fingerprint of the is1_3h_2s example on
gcc 4.8 in the debug configuration.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>