Add support for multiple apps using the HCA
For Dimemas, we will need to simulate having multiple instances of the
same MPI application on the same HCA.  This changes the HCA NED model to
allow the user to specify the number of application instances on the
same host.  Any subset of these applications can be activated at once.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>