IBOutBuf: Add pendingFreeCount
This works around a bug in our understanding of the IBOutBuf.  We cannot
simply use prevPop == DATA_FLIT to determine whether or not we need to
send a FREE message back to the VLArb, since we can't clear it since we
need to know if we are in the middle of a message or not.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>