IBOutBuf: Send flow control even if message in queue
It is possible to create a situation in which either there is a flit
in the queue or we are in the middle of a packet so we never send a flow
control message.  Fix this by sending flow control messages even when
there is an element in the queue.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>