Controller: Move delay to sending of actual data on rendezvous
This puts the delay at the beginning of the RDMA WRITE (or round-trip of
the RDMA READ) rather than before the start of the rendezvous.  The net
result is that things are pipelined better so that the round trip times
at 16KiB and 32KiB are closer to their real values.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>