Fix indentation
Some files had a random mix of spaces and tabs for indentation, and the
tab size was not consistent across files or even within a file.  Fix all
source files to not mix spaces and tabs for indentation, such that all
surrounding indentation is consistent and readable.  Source files which
exclusively used tabs for indentation were not modified.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>